Plurk: overwrite PostWidget::prepareStatus()
PostWidget::prepareStatus() is suitable for Plurk's content, so I move
the useful code from it and do the modification in
PlurkPostWidget::prepareStatus()

BTW, post->source is not the way I though, it is used to indicate the
source of this post, such as from hellotext, ping.fm, etc., not the
content_raw of the post. So, comments it out for now.

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>