Plurk: add plurk user and post data struct
Add PlurkApiUser and PlurkApiPost class to fullfill the need of Plurk.
And add a skeleton of Plurk Responses API.

The files/functions here are not useable now, we need OAuth function here
to make the Plurk APIs work.
And we also need an adaptor to transform the data structure between Choqok
and Plurk.

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>